46th Commit
Minor changes for the demo:
- Typewriter font implemented in the Book scene.
- Sound fx of Misery object fixed in Exploration scene.
- Fade transitions between doors fixed in Exploration scene.
- Folders organization.
- Countdown time set to 2 minutes for the presentation.
- Health can be also modified with keyboard arrows.

+++ Assets/Scripts/Exploration/CountdownTimer.cs
@@ -7,7 +7,7 @@ public class CountdownTimer : MonoBehaviour
     public TextMeshProUGUI timerText;
 
     [Header("Tiempo")]
-    public int startMinutes = 5;
+    public float startMinutes = 5;
 
     public LevelExit levelexit;
